requirements.txt
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
funcsigs==1.0.2
mock==2.0.0
pbr==1.10.0
peewee==2.8.2
pkg-resources==0.0.0
ptyprocess==0.5.1
simplepam==0.1.5
singledispatch==3.4.0.3
six==1.10.0
terminado==0.6
tornado==4.3
pydispatcher==2.0.5
watchdog==0.8.3
